changed tags from short to doubles; eliminated multiplication/division by PR_ALIGN_SCORE_PRECISION

+++ src/dpal.h
@@ -124,12 +124,12 @@ typedef struct {
 
 /* Structure for receiving results from the main function, dpal. */
 typedef struct {
-    const char *msg;
-    int   path[DPAL_MAX_ALIGN][2];
-    int   path_length;
-    int   align_end_1; /* Last alignment position in the 1st sequence. */
-    int   align_end_2; /* Last alignment position in the 2nd sequence. */
-    int   score;
+    const   char *msg;
+    int     path[DPAL_MAX_ALIGN][2];
+    int     path_length;
+    int     align_end_1; /* Last alignment position in the 1st sequence. */
+    int     align_end_2; /* Last alignment position in the 2nd sequence. */
+    double  score;
 } dpal_results;
 
 /* Initialize the argument to the default matrix for nucleotide matches. */
